third_party/WebKit/LayoutTests/webmidi/requestmidiaccess-expected.txt

Tests navigator.requestMIDIAccess.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.requestMIDIAccess().then(successCallback1, errorCallback1) did not throw exception.
PASS requestMIDIAccess() succeeded with access [object MIDIAccess].
PASS access.inputs().length is 1
PASS access.outputs().length is 1
PASS input attributes are correct
PASS output attributes are correct
PASS a note on message is sent without timestamp
PASS a note on message is sent with timestamp
PASS output.send([0xfff, 0x45, 0x7f]) threw exception TypeError: Failed to execute 'send' on 'MIDIOutput': The value at index 0 (4095) is greater than 0xFF..
PASS output.send([0xf0, 0x45, 0xf7]) threw exception InvalidAccessError: Failed to execute 'send' on 'MIDIOutput': System exclusive message is not allowed at index 0 (240)..
PASS requestMIDIAccess() was correctly blocked for System Exclusive access with error [object DOMError].
PASS successfullyParsed is true

TEST COMPLETE
